feat(report) Add new score value for aggregation of several modules
When using several modules, the current mutation coverage was not enough accurate. If the same mutation survived in one submodule but was killed by a test in another module (typically an integration test), the result was survived. The mutation coverage was then 0/2 (0%). Developer HAD TO analyze manually each line to be sure that the same mutation really survived.

Now developer can quickly see packages and files that need to be covered. Only the unique mutations are considered.